Android emulator & App fix
Running "npm start" first  then "npm run android" now properly runs android emulator and load App skeleton
