+++ EatMoveSleep/frontend/android/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Thu May 08 21:40:51 CEST 2025
+#Sat May 10 17:52:37 CEST 2025
 gradle.version=8.14
