+++ EatMoveSleep/frontend/package.json
@@ -15,37 +15,30 @@
   "dependencies": {
     "@react-navigation/native": "7.1.8",
     "glob": "^11.0.2",
-    "react": "19.1.0",
-    "react-native": "0.79.2",
+    "react": "18.3.1",
+    "react-native": "0.77.2",
     "react-native-gesture-handler": "2.25.0",
-    "react-native-reanimated": "3.17.5",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "4.10.0",
-    "react-native-vector-icons": "10.2.0",
     "tailwind-rn": "4.2.0"
   },
   "devDependencies": {
-    "@babel/core": "7.27.1",
-    "@babel/preset-env": "7.27.2",
-    "@babel/runtime": "7.27.1",
-    "@react-native-community/cli": "19.0.0",
-    "@react-native-community/cli-platform-android": "19.0.0",
-    "@react-native-community/cli-platform-ios": "19.0.0",
-    "@react-native/babel-preset": "0.79.2",
-    "@react-native/eslint-config": "0.79.2",
-    "@react-native/metro-config": "0.79.2",
-    "@react-native/typescript-config": "0.79.2",
-    "@types/jest": "29.5.14",
-    "@types/react": "19.1.3",
-    "@types/react-test-renderer": "19.1.0",
-    "concurrently": "9.1.2",
-    "eslint": "9.26.0",
-    "jest": "29.7.0",
-    "postcss": "8.5.3",
-    "prettier": "3.5.3",
-    "react-test-renderer": "19.1.0",
-    "tailwindcss": "4.1.5",
-    "typescript": "5.8.3"
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.3",
+    "@babel/runtime": "^7.25.0",
+    "@react-native-community/cli": "15.0.1",
+    "@react-native-community/cli-platform-android": "15.0.1",
+    "@react-native-community/cli-platform-ios": "15.0.1",
+    "@react-native/babel-preset": "0.77.2",
+    "@react-native/eslint-config": "0.77.2",
+    "@react-native/metro-config": "0.77.2",
+    "@react-native/typescript-config": "0.77.2",
+    "@types/jest": "^29.5.13",
+    "@types/react": "^18.2.6",
+    "@types/react-test-renderer": "^18.0.0",
+    "eslint": "^8.19.0",
+    "jest": "^29.6.3",
+    "prettier": "2.8.8",
+    "react-test-renderer": "18.3.1",
+    "typescript": "5.0.4"
   },
   "engines": {
     "node": ">=18"
